Test function plans moved to separate files.
