+++ Tests/Functional/Fit/Minimizer/FunctionTestPlan.cpp
@@ -13,7 +13,6 @@
 // ************************************************************************** //
 
 #include "FunctionTestPlan.h"
-#include "ObjectiveTestFunctions.h"
 #include "ParameterPlan.h"
 #include "Parameters.h"
 #include "Numeric.h"
@@ -102,27 +101,4 @@ bool FunctionTestPlan::minimumAsExpected(double found_minimum,  double tolerance
     std::cout << text.str() << std::endl;
 
     return success;
-}
-
-//! Plan to minimize a Rosenbrock function.
-//! start point: F(-1.2,1.0) = 24.20
-//! minimum    : F(1.0,1.0)  = 0.
-
-RosenbrockPlan::RosenbrockPlan() : FunctionTestPlan("RosenbrockPlan", TestFunctions::RosenBrock, 0.0)
-{
-    addParameter(Parameter("par0", -1.2, AttLimits::limited(-5.0, 5.0), 0.01), 1.0);
-    addParameter(Parameter("par1", 1.0, AttLimits::limited(-5.0, 5.0), 0.01), 1.0);
-}
-
-//! Plan for WoodFour function
-//!   start point: F(-3,-1,-3,-1) = 19192
-//!   minimum    : F(1,1,1,1)  =   0.
-
-WoodFourPlan::WoodFourPlan()
-    : FunctionTestPlan("WoodFourPlan", TestFunctions::WoodFour, 0.0)
-{
-    addParameter(Parameter("par0", -3.0, AttLimits::limited(-5.0, 5.0)), 1.0);
-    addParameter(Parameter("par1", -1.0, AttLimits::limited(-5.0, 5.0)), 1.0);
-    addParameter(Parameter("par2", -3.0, AttLimits::limited(-5.0, 5.0)), 1.0);
-    addParameter(Parameter("par3", -1.0, AttLimits::limited(-5.0, 5.0)), 1.0);
 }
